+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Features / Bug fix
 - **Custom biomes are fully supported**. Structures will be generate correctly in custom biomes !
-- **Caves biomes can be merge inside the final world** without needing an extra caves_world. And they are merged only under the world surface.
+- **Caves biomes can be merge inside the final world** without needing an extra cavesWorld. And they are merged only under the world surface.
 - **Carvers **can be run before world merging or after world merging. It is now possible to have carvers caves but still have a **clean surface without holes**. This option can be configure for each biome.
 - **Chunks outside** of the surface world are now generated **empty**.